Refactor StyledChatButton to be based on intermediate ChatButton.
ChatButton in turn, renders a Button and passes the `text` and `onClick` props through to specialize Button.
The client that uses ChatButton can now clearly see that ChatButton accepts a `chat` prop.
In the previous commit, the client saw the accepted props as `text` and `onClick`, but theses are abstracted
away by ChatButton