Update Chat e2e test
Previous attempts checked if the last message bottom distance equalled the chat window bottom distance.
This was failing from what seems like an invisible buffer of around 10 pixels that the browser adds to
the container of the scrollbar.

The last message was still visually touching the bottom of the chat viewport.

This change asserts the last message is at the bottom by:
- checking that the last message is visible in the chat viewport, and
- checking that the chat is fully scrolled to the bottom